Interfaces Exercises completed2

+++ src/org/launchcode/java/demos/lsn7interfaces/ConeComparator.java
@@ -6,7 +6,7 @@ public class ConeComparator implements Comparator<Cone> {
 
     @Override
     public int compare(Cone o1, Cone o2) {
-        return o1.getCost() - (o2.getCost());
+        return (int) (o1.getCost() - (o2.getCost()));
 
 
         //TODO: need help with this.
